backend/requirements.txt
livekit
livekit-agents
livekit-plugins-deepgram
livekit-plugins-groq
fastapi
uvicorn
python-dotenv
